second commit

+++ Type casting.py
@@ -8,7 +8,7 @@ celsius_value = input("temperature value in celsius = ")
 
 print ((int(celsius_value) * 1.8) + 32)
 
-print (int(celsius_value) * (9/5) + 32)
+print (float(celsius_value) * (9/5) + 32)
 
 fahrenheit_value = input("final answer in fahrenheit is = ")
